+++ pcf/particle/aws/batch/batch_job_queue.py
@@ -34,6 +34,12 @@ class BatchJobQueue(AWSResource):
         "INVALID": State.terminated,
     }
 
+    UPDATE_PARAM_FILTER = {
+        "state",
+        "priority",
+        "computeEnvironmentOrder",
+    }
+
     UNIQUE_KEYS = ['aws_resource.jobQueueName']
 
     def __init__(self, particle_definition, session=None):
@@ -43,45 +49,47 @@ class BatchJobQueue(AWSResource):
 
     def _set_unique_keys(self):
         """
-        Logic that sets keys from state definition that are used to uniquely identify the BatchComputeEnvironment
+        Logic that sets keys from state definition that are used to uniquely identify the BatchJobQueue
         """
         self.unique_keys = BatchJobQueue.UNIQUE_KEYS
 
     def get_status(self):
         """
-        Calls boto3 describe_compute_environments.
+        Calls boto3 describe_job_queues.
 
         Returns:
              status or {}
         """
-        res = self.client.jobQueues(
+        res = self.client.describe_job_queues(
             jobQueues=[self.name],
         )
 
-        if len(res["computeEnvironments"]) != 1:
+        if len(res["jobQueues"]) != 1:
             return {}
 
-        return res["computeEnvironments"][0]
+        return res["jobQueues"][0]
 
     def _terminate(self):
         """
-        Sets state DISABLED then calls boto3 delete_compute_environment() to terminate
+        Sets state DISABLED then calls boto3 delete_job_queue() to terminate
 
         Returns:
-            boto3 delete_compute_environment() response
+            boto3 delete_job_queue() response
         """
         # need to disable before terminating
-        self._stop()
+        if self.state != State.stopped:
+            return self._stop()
+
         return self.client.delete_job_queue(
             jobQueue=self.name
         )
 
     def _start(self):
         """
-        Calls boto3 create_compute_environment() and sets the status to ENABLED by default.
+        Calls boto3 create_job_queue() and sets the status to ENABLED by default.
 
         Returns:
-            boto3 create_compute_environment() response
+            boto3 create_job_queue() response
         """
         # calls enable if transitioning from stopped state
         if self.current_state_definition:
@@ -90,10 +98,10 @@ class BatchJobQueue(AWSResource):
 
     def enable(self):
         """
-        Calls boto3 update_compute_environment() and sets state to ENABLED
+        Calls boto3 update_job_queue() and sets state to ENABLED
 
         Returns:
-            boto3 update_compute_environment() response
+            boto3 update_job_queue() response
         """
         return self.client.update_job_queue(
             jobQueue=self.name,
@@ -102,7 +110,7 @@ class BatchJobQueue(AWSResource):
 
     def _stop(self):
         """
-        Calls boto3 update_compute_environment() and sets state to DISABLED
+        Calls boto3 update_job_queue() and sets state to DISABLED
 
         Returns:
             boto3 update_compute_environment() response
@@ -135,14 +143,15 @@ class BatchJobQueue(AWSResource):
 
     def _update(self):
         """
-        Calls boto3 update_compute_environment(). Can only update fields in UPDATE_PARAM_FILTER.
+        Calls boto3 update_job_queue(). Can only update fields in UPDATE_PARAM_FILTER.
 
         Returns:
             boto3 put_attributes() response
         """
+        filtered_definition = pcf_util.param_filter(self.desired_state_definition, self.UPDATE_PARAM_FILTER)
         return self.client.update_job_queue(
             jobQueue=self.name,
-            **self.desired_state_definition
+            **filtered_definition
         )
 
     def is_state_definition_equivalent(self):
@@ -153,8 +162,10 @@ class BatchJobQueue(AWSResource):
         Returns:
             bool
         """
+        filtered_desired_def = pcf_util.param_filter(self.desired_state_definition, self.UPDATE_PARAM_FILTER)
+        filtered_current_def = pcf_util.param_filter(self.current_state_definition, self.UPDATE_PARAM_FILTER)
 
-        diff = pcf_util.diff_dict(self.desired_state_definition, self.current_state_definition)
+        diff = pcf_util.diff_dict(filtered_desired_def, filtered_current_def)
 
         if diff or len(diff) != 0:
             return False
